+++ tests/e2e/utils/constants.py
@@ -9,4 +9,4 @@ DEFAULT_USERNAME = "user@example.com"
 DEFAULT_PASSWORD = "12341234"
 KUBEFLOW_GROUP = "kubeflow.org"
 KUBEFLOW_NAMESPACE = "kubeflow"
-KUBEFLOW_VERSION = "v1.4.1"
+KUBEFLOW_VERSION = "v1.5.0"
