+++ tests/e2e/utils/k8s_core_api.py
@@ -2,6 +2,8 @@
 Module for helper methods to create and delete kubernetes core api resources (e.g. pods, namespaces, etc.)
 """
 
+import subprocess
+
 from kubernetes.client import V1Namespace
 from kubernetes.client.exceptions import ApiException
 
@@ -15,3 +17,16 @@ def create_namespace(cluster, region, namespace_name):
     except ApiException as e:
         if "Conflict" != e.reason:
             raise e
+
+def upload_file_as_configmap(namespace, configmap_name, file_path):
+    subprocess.call(
+        f"kubectl create configmap -n {namespace} {configmap_name} --from-file {file_path}".split()
+    )
+
+def delete_configmap(namespace, configmap_name):
+    subprocess.call(f"kubectl delete configmap -n {namespace} {configmap_name}".split())
+
+def patch_configmap(namespace, configmap_name, patch_file_path):
+    subprocess.call(
+        f"kubectl patch configmap -n {namespace} {configmap_name} --patch-file {patch_file_path}".split()
+    )
